Update tag via bash instead of frau-ci
frau-ci was having troubles with the approach we're using here as it doesn't first pull the most recent stuff before trying to push. This means it's missing the Travis commit. So, figured might as well just do what frau-ci is doing, but here. Only missing part is updating the `version` in package.json, but that doesn't have much of an effect when we're distributing this via bower, not npm.