Use deumdify browserify plugin
We were seeing an issue where requirejs being on the same page was causing problems, specifically http://requirejs.org/docs/errors.html#mismatch. This was only happening on UMD FRA pages in the LMS. We're not using the UMD-ness of d2l-fetch, so might as well remove it to stop causing problems.

Note that this will be a major version bump since this is technically a breaking change (albeit breaking in a way we don't use, but better safe than sorry).